move version tests
